InteractionHistory

Description

This module handles the interaction history of the Haskell-Based Social Network Simulation. It includes functionalities to connect to a SQLite database, log interactions, and query the interaction history.

Documentation

connectDb :: IO Connection

Connects to the SQLite database used for storing interaction history.

initDb :: Connection -> IO ()

Initializes the database by creating a table for messages if it does not exist.

logInteraction :: Connection -> Message -> IO ()

Logs a new interaction (message) in the database.

param conn The database connection. param message The message to log.

queryUserHistory :: Connection -> Int -> Int -> IO [Message]

Queries the interaction history of a specific user.

param conn The database connection. param senderUserId The user ID of the sender. param receiverUserId The user ID of the receiver. return A list of messages related to the specified user.

queryFullHistory :: Connection -> IO [Message]

Queries the entire interaction history.

param conn The database connection. return A list of all messages in the interaction history.