Main

Description

Documentation

main :: IO ()

Main functions that initiliase users, threads and invoke the other main functions

process :: User -> MVar Int -> MVar [Message] -> MVar [Message] -> [User] -> IO ()

Create a new thread that allow a user to send a message to random user The user must have at least one common interest The process terminate once all the processes have sent 100 messages in total Delay is set 1-5 seconds

logginUser :: [User] -> [Message] -> IO ()

Login the user Check credentials Invoke menu that allow the user to check messaging history

main_menu :: [Message] -> String -> User -> IO ()

Main menu that allow logged in user to see his/her conversations

countOccurences :: Ord a => [a] -> [(a, Int)]

Used to count the number of occurence, in this case the number of messages received for each user

atRandIndex :: [a] -> IO a

Used to select a random item from a list

generateRandomDate :: IO Day

Pick a random date between 20220101 - 20230101

generateRandomMessage :: IO String

Pick a random message from the list

printInformation :: (String, Int) -> String

Used assists printing when counting the number of messages received by each user.

printChatHistory :: Message -> String -> String

Used to print chat history between the logged in user and the others

exists :: Eq a => [a] -> [a] -> Bool

Used to check if an item of one list contain at least one is present in other list

filterOutCurrentUser :: User -> [User] -> [User]

Used to filter out current users and return other users, and help to prevent self-messaging

showNumberOfReceivedMessages :: [Message] -> IO ()

Display the number of messages received for each users

showHistory :: [Message] -> String -> IO ()

Show between the current user and all other users

showConversationBetweenTwo :: [Message] -> String -> String -> IO ()

Show a conversation between two users

checkIfSameInterests :: User -> User -> IO Bool

Check if two users have at least one common interest