Message

Description

Module for handling messages in the Haskell-Based Social Network Simulation. This module defines the Message data type and provides functionalities for creating and managing messages.

Documentation

data Message

Represents a message in the social network simulation. A message consists of a sender, a receiver, and the message content.

Constructors

Message

Fields

sender :: User

The user who sends the message.

receiver :: User

The user who receives the message.

content :: String

The content of the message.

createMessage :: User -> User -> String -> Message

Creates a new message with specified sender, receiver, and content.

param from The sender of the message. param to The receiver of the message. param msgContent The content of the message. return The constructed Message.

randomMessageContent :: IO String

Generates a random message content from a predefined list of phrases.

@return A randomly selected message content.

createRandomMessage :: User -> User -> IO Message

Creates a random message between two users.

This function selects a random content and constructs a message from one user to another.

param from The sender of the message. param to The receiver of the message. @return A randomly generated Message in IO context.