Types

Description

Module for defining core data types used in the Haskell-Based Social Network Simulation. This includes types for representing users and messages in the network.

Documentation

data User

Represents a user in the social network simulation. Each user is characterized by a unique identifier and a username.

Constructors

User

Fields

userId :: Int

The unique identifier of the user.

username :: String

The username of the user.

data Message

Represents a message in the social network simulation. A message consists of a sender, a receiver, and the message content.

Constructors

Message

Fields

sender :: User

The user who sends the message.

receiver :: User

The user who receives the message.

content :: String

The content of the message.