User

Description

Module for handling user-related functionalities in the Haskell-Based Social Network Simulation. This module provides functionalities for creating and managing users, including generating random users for the simulation.

Documentation

createUser :: Int -> String -> User

Creates a new user with a specified ID and username.

param id The unique identifier for the new user. param name The name for the new user. @return A new User object.

randomUsername :: IO String

Generates a random username from a predefined list of names.

@return A randomly selected username.

createRandomUser :: IO User

Creates a random user with a unique ID and a randomly generated username.

@return A randomly generated User in IO context.

getUsername :: User -> String

Retrieves the username of a given User.

param user The User object. return The username of the User.